server/auth.py
from passlib.context import CryptContext
from jose import jwt
from datetime import datetime, timedelta
import os

# Password hashing context
pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")

# JWT Configuration
ALGORITHM = "RS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 1440  # 24 hours

# Fix path resolution
BASE_DIR = os.path.dirname(os.path.abspath(__file__))
PRIVATE_KEY_PATH = os.path.join(BASE_DIR, "private_key.pem")

def get_private_key():
    # In production, you might want to load this from an Env Var for better security
    # But for now, we load from file
    with open(PRIVATE_KEY_PATH, "rb") as f:
        return f.read()

def get_public_key():
    public_key_path = os.path.join(BASE_DIR, "public_key.pem")
    with open(public_key_path, "rb") as f:
        return f.read()

def verify_password(plain_password, hashed_password):
    return pwd_context.verify(plain_password, hashed_password)

def get_password_hash(password):
    return pwd_context.hash(password)

def create_access_token(data: dict, expires_delta: timedelta | None = None):
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=15)
    
    to_encode.update({"exp": expire})
    
    private_key = get_private_key()
    encoded_jwt = jwt.encode(to_encode, private_key, algorithm=ALGORITHM)
    return encoded_jwt, expire
